airflow/requirements.txt
SQLAlchemy==1.4.46
click==8.1.3
ruamel.yaml==0.17.21
python-dotenv==0.21.1
dataflows==0.3.16
requests==2.28.2
beautifulsoup4==4.11.2
google-cloud-storage==2.7.0
